fix: sonar issues
* feat: cretate reusable icon button with tooltip

* fix: account edition layout sonar lint

* fix: account show sonar issues
