+++ src/common/components/index.ts
@@ -6,6 +6,7 @@ export * from './BpFormField';
 export * from './BPLayout';
 export * from './BPLoader';
 export * from './BpMultipleTextInput';
+export * from './Buttons';
 export * from './EmptyListTemplate';
 export * from './Field';
 export * from './Pdf';
