+++ src/operations/account/components/index.ts
@@ -1,1 +1,11 @@
+export * from './AccountEditionAccordion';
+export * from './AccountHolderIncomeTargets';
+export * from './AccountHolderLayout';
+export * from './AdditionalInformation';
 export * from './FeedbackLink';
+export * from './InfoShow';
+export * from './LogoShowLayout';
+export * from './ProfileShowLayout';
+export * from './SubjectToVatSwitch';
+export * from './SubscriptionLayout';
+export * from './types';
